+++ CMakeLists.txt
@@ -91,14 +91,15 @@ set(XEUS_SQLITE_SRC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/src)
 set(XEUS_SQLITE_SRC
     ${XEUS_SQLITE_SRC_DIR}/xeus_sqlite_interpreter.cpp
     ${XEUS_SQLITE_SRC_DIR}/xvega_bindings.cpp
+    ${XEUS_SQLITE_SRC_DIR}/xparser.hpp
     ${XEUS_SQLITE_SRC_DIR}/utils.cpp
+    ${XEUS_SQLITE_SRC_DIR}/utils.hpp
 )
 
 set(XEUS_SQLITE_HEADERS
     include/xeus-sqlite/xeus_sqlite_interpreter.hpp
     include/xeus-sqlite/xvega_bindings.hpp
     include/xeus-sqlite/xeus_sqlite_config.hpp
-    include/xeus-sqlite/utils.hpp
 )
 
 set(XSQLITE_SRC ${XEUS_SQLITE_SRC_DIR}/main.cpp)
