+++ include/xeus-sqlite/xeus_sqlite_config.hpp
@@ -12,7 +12,7 @@
 
 // Project version
 #define XSQLITE_VERSION_MAJOR 0
-#define XSQLITE_VERSION_MINOR 1 
+#define XSQLITE_VERSION_MINOR 1
 #define XSQLITE_VERSION_PATCH 1
 
 // Composing the version string from major, minor and patch
